Delete custom logging to blob storage and webhook
