+++ packages/publisher/package.json
@@ -15,7 +15,6 @@
     "adal-node": "^0.1.22",
     "applicationinsights": "^1.0.7",
     "azure-keyvault": "^3.0.4",
-    "azure-storage": "^2.0.0",
     "fs-extra": "4.0.0",
     "fstream": "^1.0.12",
     "longjohn": "^0.2.11",
@@ -48,12 +47,8 @@
     "publish-packages": "node -r source-map-support/register dist/publish-packages.js",
     "publish-packages-dry": "node -r source-map-support/register dist/publish-packages.js --dry",
     "publish-registry": "node -r source-map-support/register dist/publish-registry.js",
-    "upload-blobs": "node -r source-map-support/register dist/upload-blobs.js",
     "full": "node -r source-map-support/register dist/full.js",
     "full-dry": "node -r source-map-support/register dist/full.js --dry",
-    "webhook-dry": "node -r source-map-support/register dist/webhook.js --dry",
-    "make-server-run": "node -r source-map-support/register dist/make-server-run.js",
-    "make-production-server-run": "node -r source-map-support/register dist/make-server-run.js --remote",
     "test-tsNext": "node -r source-map-support/register dist/tester/test.js --all --tsNext"
   },
   "repository": {
