+++ packages/publisher/src/lib/secrets.ts
@@ -15,14 +15,6 @@ export enum Secret {
    */
   GITHUB_ACCESS_TOKEN,
   /**
-   * This is used to ensure that only GitHub can send messages to our server.
-   * This should match the secret value set on GitHub: https://github.com/DefinitelyTyped/DefinitelyTyped/settings/hooks
-   * The Payload URL should be the URL of the Azure service.
-   * The webhook ignores the `sourceRepository` setting and can be triggered by *anything* with the secret,
-   * so make sure only DefinitelyTyped has the secret.
-   */
-  GITHUB_SECRET,
-  /**
    * Token used to perform request to NPM's API.
    * This was generated by doing:
    * - `npm login`
