+++ packages/publisher/src/lib/settings.ts
@@ -10,12 +10,5 @@ export const logDir = joinPaths(root, "logs");
 export const definitelyTypedZipUrl = "https://codeload.github.com/DefinitelyTyped/DefinitelyTyped/tar.gz/master";
 /** The branch that DefinitelyTyped is sourced from. */
 export const sourceBranch = "master";
-
-/** Name of the azure storage account. Used for uploading data and logs. */
-export const azureStorageAccount = "typespublisher";
-/** Name of the azure container. */
-export const azureContainer = "typespublisher";
 /** URL of azure keyvault. */
 export const azureKeyvault = "https://types-publisher-keys.vault.azure.net";
-/** Issue in types-publisher that we will use to report webhook errors. */
-export const errorsIssue = "Microsoft/types-publisher/issues/40";
