--- packages/publisher/src/upload-blobs.ts
@@ -1,9 +1,0 @@
-import uploadBlobsAndUpdateIssue from "./lib/blob-uploader";
-import { currentTimeStamp } from "./util/util";
-import { logUncaughtErrors } from "@definitelytyped/utils";
-
-if (!module.parent) {
-  logUncaughtErrors(uploadBlobsAndUpdateIssue(currentTimeStamp()));
-}
-
-export default uploadBlobsAndUpdateIssue;
